authorization api and jwt created
using Django rest framework authorization api and jwt access and refresh token created.
To register use register api (it will send an otp to the mail of the user who wants to register). verify api - to verify the otp, forgot api - it will send otp to users mail to verify the mail for the forgot account and reset api - mail, otp and new_password to reset the pass.

Jwt
token api - to get both refresh and access token of a user
refresh token - to get the access token by sending refresh token